Use my version of projectile-rails and add project specific inits
As the comment says, setting "risky" in .dir-locals.el is painful
because you must approve them for each file.  Instead, I created a
small routine and call it via eval.  This seems safe to me and
prevents Emacs from pestering me so much.

I also discovered that .dir-locals.el is NOT loaded when dired is run
on the directory.  A file must be loaded.

Last, I pulled down my branch of projectile-rails that adds buttons to
errors shown in the server log when running the server in a projectile
window.