Changed comments to no
I updated to Emacs 30 and had a problem with loading this file.  Emacs
complained that lexical-binding needed to be set on the first line.  I
don't know why.  So I tinkered around until I got it to work by
changing the first line from `:comments link` to `:comments no` and
adding the code block to add the lexical binding comment.